+++ src/Infrastructure/Http/index.ts
@@ -1,7 +1,9 @@
-export * from './AbstractController';
 export * from './RestHelperFunctions';
 export * from './HttpOrderedInterceptorGroup';
 export * from './UncaughtErrorCatcher';
 export * from './RedirectResult';
-export * from './HcHttpModule';
+
 export * from './Bootstrap';
+export * from './Controller';
+
+export * from './HcHttpModule';