+++ src/Infrastructure/index.ts
@@ -1,6 +1,4 @@
 export * from './Http';
 export * from './Persistence/Domain';
-export * from './Redis';
 export * from './Config';
-
 export * from './Account';